// src/common/enums/status.enum.ts
export enum Status {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  IGNORED = 'IGNORED',
  INVOICE = 'INVOICE',
  QUOTATION = 'QUOTATION',
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  EXPIRED = 'EXPIRED',
  NEW = 'NEW',
  VIEWED = 'VIEWED',
  SENT = 'SENT',
  FAILED = 'FAILED',
  QUEUED = 'QUEUED',
}